health-food/pyproject.toml
[project]
name = "health-food"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.5",
    "httpx>=0.28.1",
    "jinja2==3.1.4",
    "openai>=1.100.2",
    "openai-agents>=0.2.8",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]==2.11.7",
    "pydantic-core==2.33.2",
    "python-dotenv==1.0.1",
    "python-jose>=3.5.0",
    "python-multipart==0.0.12",
    "starlette==0.41.3",
    "uvicorn==0.30.6",
]
